change MainPage  Module for new Usecases (Fragment,ViewModel and ViewModelFactory)

+++ MainPage/src/main/java/ir/omidtaheri/mainpage/entity/GenreUiEntity.kt
@@ -1,3 +1,3 @@
 package ir.omidtaheri.mainpage.entity
 
-data class GenreUiEntity(val id: Long, val name: String)
+data class GenreUiEntity(val id: Int, val name: String)
